ui: fix linting issues after merge
