+++ ui/src/pages/Channel.tsx
@@ -1,20 +1,11 @@
-import _ from 'lodash';
 import React from 'react';
 import { useParams } from 'react-router';
-import ChatInput from '../components/ChatInput/ChatInput';
 import ChatWindow from '../components/ChatWindow';
-import { useChatIsJoined, useChatPerms, useChatState } from '../state/chat';
-import { useRouteGroup, useVessel } from '../state/groups';
+import { useChatIsJoined, useChatState } from '../state/chat';
 
 function Channel() {
   const { chShip, chName } = useParams();
   const flag = `${chShip}/${chName}`;
-  const groupFlag = useRouteGroup();
-  const perms = useChatPerms(flag);
-  const vessel = useVessel(groupFlag, window.our);
-  const canWrite =
-    perms.writers.length === 0 ||
-    _.intersection(perms.writers, vessel.sects).length !== 0;
   const isJoined = useChatIsJoined(flag);
   const join = () => {
     useChatState.getState().joinChat(flag);
